Create dummy start_wireguard for non-linux
