+++ common/wireguard/Cargo.toml
@@ -28,5 +28,5 @@ tap.workspace = true
 thiserror.workspace = true
 tokio = { workspace = true, features = ["rt-multi-thread", "net", "io-util"] }
 
-[target.'cfg(target_os = "linux")'.dependencies] 
+[target.'cfg(target_os = "linux")'.dependencies]
 tokio-tun = "0.9.0"
